View CSR modal has code duplicated with src/app/certificate-authorities/
